I altered functions.
Now my library compiles completely. However, 9 functions of part one do not work. 14 functions do work and pass the test. When these 9 functions work, I will have passed part 1.

+++ Libft_remote/ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: lhop <lhop@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/11 13:08:45 by lhop          #+#    #+#                 */
-/*   Updated: 2022/10/24 13:32:26 by lhop          ########   odam.nl         */
+/*   Updated: 2022/10/28 14:36:16 by lhop          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *string)
+#include "libft.h"
+
+size_t	ft_strlen(char *string)
 {
 	int	i;
