+++ Libft_remote/libft.h
@@ -6,7 +6,7 @@
 /*   By: lhop <lhop@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/20 14:27:09 by lhop          #+#    #+#                 */
-/*   Updated: 2022/10/27 11:43:30 by lhop          ########   odam.nl         */
+/*   Updated: 2022/10/28 14:35:52 by lhop          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int		ft_isdigit(int c);
 int		ft_isalnum(int c);
 int		ft_isascii(int c);
 int		ft_isprint(int c);
-int		ft_strlen(char *string);
+size_t	ft_strlen(char *string);
 char	*ft_strnstr(const char *hay, const char *needle, size_t len);
 size_t	ft_strlcat(const char *src, char *dst, size_t dstsize);
 size_t	ft_strlcpy(char *dst, char *src, size_t dstsize);
